Move rewrite_mod to the items module
