+++ rustfmt-core/src/imports.rs
@@ -124,16 +124,6 @@ pub fn rewrite_import(
     }
 }
 
-/// Rewrite an inline mod.
-pub fn rewrite_mod(item: &ast::Item) -> String {
-    let mut result = String::with_capacity(32);
-    result.push_str(&*format_visibility(&item.vis));
-    result.push_str("mod ");
-    result.push_str(&item.ident.to_string());
-    result.push(';');
-    result
-}
-
 impl<'a> FmtVisitor<'a> {
     pub fn format_imports(&mut self, use_items: &[&ast::Item]) {
         if use_items.is_empty() {
